pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "awesome-audio"
version = "0.4.0"
description = "A curated guide to open-source audio and music projects"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
]

[project.scripts]
awesome-audio = "src.cli:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "ruff>=0.14.14",
    "sqlalchemy>=2.0.46",
]
